+++ src/main/java/roomescape/controller/api/MemberRestController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
